# src/bkauth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BKAuth API 文档",
        "title": "BKAuth API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/apps/{bk_app_code}/access-keys": {
            "get": {
                "description": "list app secret",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app secret list",
                "operationId": "api-app-secret-list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the app which want to list secret",
                        "name": "bk_app_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.AccessKeyWithCreatedAt"
                                            }
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "creates app secret",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app secret create",
                "operationId": "api-app-secret-create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the app which want to create secret",
                        "name": "bk_app_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.AccessKey"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{bk_app_code}/access-keys/verify": {
            "post": {
                "description": "verify app secret",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app secret verify",
                "operationId": "api-app-secret-verify",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the app which want to verify secret",
                        "name": "bk_app_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "app secret",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.appSecretSerializer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/apps/{bk_app_code}/access-keys/{access_key_id}": {
            "put": {
                "description": "put app secret",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app secret put",
                "operationId": "api-app-secret-put",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the app which want to put secret",
                        "name": "bk_app_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the secret which want to delete",
                        "name": "access_key_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "app secret",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.accessKeyUpdateSerializer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "delete app secret",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "app"
                ],
                "summary": "app secret delete",
                "operationId": "api-app-secret-delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app_code",
                        "name": "X-BK-APP-CODE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "app_secret",
                        "name": "X-BK-APP-SECRET",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the app which want to delete secret",
                        "name": "bk_app_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the secret which want to delete",
                        "name": "access_key_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.Response"
                        },
                        "headers": {
                            "X-Request-Id": {
                                "type": "string",
                                "description": "the request id"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.accessKeyUpdateSerializer": {
            "type": "object",
            "required": [
                "enabled"
            ],
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handler.appSecretSerializer": {
            "type": "object",
            "required": [
                "bk_app_secret"
            ],
            "properties": {
                "bk_app_secret": {
                    "type": "string",
                    "maxLength": 128,
                    "example": "bk_paas"
                }
            }
        },
...
        "types.AccessKey": {
            "type": "object",
            "properties": {
                "bk_app_code": {
                    "type": "string"
                },
                "bk_app_secret": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "types.AccessKeyWithCreatedAt": {
            "type": "object",
            "properties": {
                "bk_app_code": {
                    "type": "string"
                },
                "bk_app_secret": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "util.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}